requirements.txt
pyodbc>=4.0.39
requests>=2.31.0
python-dateutil>=2.8.2
pyinstaller>=6.3.0
